Force user to authenticate again if they remove the only account in account manager

+++ app/src/main/java/xyz/hisname/fireflyiii/ui/HomeActivity.kt
@@ -86,15 +86,22 @@ class HomeActivity: BaseActivity(){
             startActivity(Intent(this, AuthActivity::class.java))
             finish()
         } else {
-            setContentView(view)
-            homeViewModel.migrateFirefly()
-            setup()
-            if(keyguardUtil.isAppKeyguardEnabled()){
-                binding.biggerFragmentContainer.isInvisible = true
-                authenticator = Authenticator(this, ::handleResult)
-                authenticator.authenticate()
+            homeViewModel.accountDoesNotExistInAccountManager().observe(this){ exist ->
+                if(exist){
+                    startActivity(Intent(this, AuthActivity::class.java))
+                    finish()
+                } else {
+                    setContentView(view)
+                    homeViewModel.migrateFirefly()
+                    setup()
+                    if(keyguardUtil.isAppKeyguardEnabled()){
+                        binding.biggerFragmentContainer.isInvisible = true
+                        authenticator = Authenticator(this, ::handleResult)
+                        authenticator.authenticate()
+                    }
+                    homeViewModel.deleteUnusedAccount()
+                }
             }
-            homeViewModel.deleteUnusedAccount()
         }
     }
